requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
httpx>=0.27.0
rich>=13.0.0
